Fix power bi create group

+++ MillimanAccessPortal/MillimanAccessPortal/Controllers/ContentPublishingController.cs
@@ -517,7 +517,14 @@ namespace MillimanAccessPortal.Controllers
                     PowerBiContentItemProperties props = rootContentItem.TypeSpecificDetailObject as PowerBiContentItemProperties;
 
                     PowerBiLibApi powerBiApi = await new PowerBiLibApi(_powerBiConfig).InitializeAsync();
-                    await powerBiApi.DeleteReportAsync(props.LiveReportId.Value);
+                    if (props is not null && props.LiveReportId.HasValue)
+                    {
+                        await powerBiApi.DeleteReportAsync(props.LiveReportId.Value);
+                    }
+                    if (props is not null && props.PreviewReportId.HasValue)
+                    {
+                        await powerBiApi.DeleteReportAsync(props.PreviewReportId.Value);
+                    }
                     break;
 
                 case ContentTypeEnum.Html:
